+++ internal/util-scripted/src/main/scala/sbt/internal/scripted/FileCommands.scala
@@ -9,7 +9,7 @@ import java.io.File
 import sbt.io.{ IO, Path }
 import sbt.io.syntax._
 import Path._
-import sbt.io.Milli.getModifiedTime
+import sbt.io.IO.getModifiedTime
 
 class FileCommands(baseDirectory: File) extends BasicStatementHandler {
   lazy val commands = commandMap
@@ -68,8 +68,7 @@ class FileCommands(baseDirectory: File) extends BasicStatementHandler {
   def newer(a: String, b: String): Unit = {
     val pathA = fromString(a)
     val pathB = fromString(b)
-    val isNewer = pathA.exists && (!pathB.exists || getModifiedTime(pathA) > getModifiedTime(
-      pathB))
+    val isNewer = pathA.exists && (!pathB.exists || getModifiedTime(pathA) > getModifiedTime(pathB))
     if (!isNewer) {
       scriptError(s"$pathA is not newer than $pathB")
     }
